# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user": {
      "post": {
        "description": "Jammit 계정 생성",
        "operationId": "JammitUserController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "email": "user@example.com",
                  "password": "password123",
                  "username": "jammituser",
                  "nickname": "닉네임",
                  "preferredGenres": [
                    "ROCK",
                    "POP"
                  ],
                  "preferredBandSessions": [
                    "VOCAL",
                    "ELECTRIC_GUITAR"
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "가입 성공 (CommonResponseDto)"
          },
          "400": {
            "description": "이메일 중복 등"
          }
        },
        "summary": "회원가입",
        "tags": [
          "Jammit User"
        ]
      },
      "get": {
        "description": "JWT 기준 로그인 유저 정보",
        "operationId": "JammitUserController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "프로필 객체 (CommonResponseDto)"
          },
          "401": {
            "description": "미인증"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "내 프로필 조회",
        "tags": [
          "Jammit User"
        ]
      },
      "put": {
        "operationId": "JammitUserController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "email": "new@example.com",
                  "username": "newname",
                  "password": "newpassword",
                  "preferredGenres": [
                    "JAZZ"
                  ],
                  "preferredBandSessions": [
                    "DRUM"
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정 후 프로필 (CommonResponseDto)"
          },
          "400": {
            "description": "유저 없음 등"
          },
          "401": {
            "description": "미인증"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "내 프로필 수정",
        "tags": [
          "Jammit User"
        ]
      }
    },
...
    "/video": {
      "get": {
        "operationId": "VideoController_getVideosWithPagination",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "페이지 (1부터)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": true,
            "in": "query",
            "description": "페이지 크기",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "order",
            "required": true,
            "in": "query",
            "schema": {
              "enum": [
                "latest",
                "popular"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "페이지네이션된 영상 목록",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VideoListFlatDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "목록 조회 실패"
          }
        },
        "summary": "전체 영상 목록",
        "tags": [
          "Video"
        ]
      }
    },
    "/video/user": {
      "get": {
        "operationId": "VideoController_getMyVideos",
        "parameters": [
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Jammit users.id — 없으면 토큰 유저 본인",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": true,
            "in": "query",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "order",
            "required": true,
            "in": "query",
            "schema": {
              "enum": [
                "latest",
                "popular"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "해당 유저가 올린 영상",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoListFlatDto"
                }
              }
            }
          },
          "401": {
            "description": "미인증"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "유저별 업로드 영상 목록",
        "tags": [
          "Video"
        ]
      }
    },
    "/video/user/count": {
      "get": {
        "operationId": "VideoController_getMyVideoTotal",
        "parameters": [
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Jammit users.id — 없으면 본인",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ count, message }",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "count": 12,
                    "message": "업로드한 영상 개수를 조회했습니다."
                  }
                }
              }
            }
          },
          "401": {
            "description": "미인증"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "유저별 업로드 영상 개수",
        "tags": [
          "Video"
        ]
      }
    },
    "/video/status": {
      "get": {
        "description": "uploadId 기준 playback 준비 여부",
        "operationId": "VideoController_getMuxStatus",
        "parameters": [
          {
            "name": "uploadId",
            "required": true,
            "in": "query",
            "schema": {
              "example": "abc-upload-id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "playbackId": "xxxxx",
                    "status": "ready"
                  }
                }
              }
            }
          }
        },
        "summary": "Mux 업로드 처리 상태",
        "tags": [
          "Video"
        ]
      }
    },
    "/video/{videoId}": {
      "get": {
        "operationId": "VideoController_getVideosWithDetail",
        "parameters": [
          {
            "name": "videoId",
            "required": true,
            "in": "path",
            "description": "영상 UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "상세 JSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "영상 없음"
          },
          "500": {
            "description": "조회 오류"
          }
        },
        "summary": "영상 상세",
        "tags": [
          "Video"
        ]
      },
      "post": {
        "description": "로그인 시 회원 UUID 기준, 비로그인 시 IP. Authorization 헤더는 선택.",
        "operationId": "VideoController_viewCountVideos",
        "parameters": [
          {
            "name": "videoId",
            "required": true,
            "in": "path",
            "description": "영상 UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-forwarded-for",
            "in": "header",
            "description": "프록시 뒤 비회원 IP (선택)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "조회수 처리 완료",
                    "videoId": "3520e5b7-af1b-405c-af5c-b0511635c6fa",
                    "viewCount": 42
                  }
                }
              }
            }
          },
          "400": {
            "description": "유저·IP 식별 불가"
          },
          "404": {
            "description": "영상 없음"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "조회수 증가",
        "tags": [
          "Video"
        ]
      },
      "delete": {
        "description": "본인 업로드만",
        "operationId": "VideoController_deleteVideo",
        "parameters": [
          {
            "name": "videoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "영상이 삭제되었습니다."
                  }
                }
              }
            }
          },
          "401": {
            "description": "미인증"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "영상 없음"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "영상 삭제",
        "tags": [
          "Video"
        ]
      }
    },
    "/video/like/{videoId}": {
      "post": {
        "operationId": "VideoController_toggleLike",
        "parameters": [
          {
            "name": "videoId",
            "required": true,
            "in": "path",
            "description": "영상 UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "좋아요/취소 결과",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "좋아요",
                    "liked": true,
                    "likeCount": 5
                  }
                }
              }
            }
          },
          "401": {
            "description": "미인증"
          },
          "404": {
            "description": "영상 없음"
          },
          "500": {
            "description": "처리 오류"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "좋아요 토글",
        "tags": [
          "Video"
        ]
      }
    },
    "/video/like-status/{videoId}": {
      "get": {
        "operationId": "VideoController_getLikeStatus",
        "parameters": [
          {
            "name": "videoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "liked": true,
                    "likeCount": 10
                  }
                }
              }
            }
          },
          "401": {
            "description": "미인증"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "좋아요 여부·개수",
        "tags": [
          "Video"
        ]
      }
    },
    "/video/uploadUrl": {
      "post": {
        "operationId": "VideoController_getUploadUrl",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MuxUploadResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "미인증"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Mux Direct Upload URL 발급",
        "tags": [
          "Video"
        ]
      }
    },
    "/video/register": {
      "post": {
        "description": "Mux 메타와 제목·설명 등 저장",
        "operationId": "VideoController_registerMuxVideo",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterMuxVideoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "등록 결과"
          },
          "401": {
            "description": "미인증"
          },
          "500": {
            "description": "등록 실패"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "업로드 완료 후 영상 등록",
        "tags": [
          "Video"
        ]
      }
    },
    "/video/webhook/mux": {
      "post": {
        "description": "Raw JSON body + Mux-Signature. Swagger에서는 서명 검증이 어려우므로 실제는 Mux/Postman 등으로 호출.",
        "operationId": "VideoController_muxWebHook",
        "parameters": [],
        "responses": {
          "200": {
            "description": "{ received: true }"
          },
          "403": {
            "description": "서명 불일치"
          }
        },
        "summary": "Mux 웹훅",
        "tags": [
          "Video"
        ]
      }
    },
...
  },
  "info": {
    "title": "jammit api문서",
    "description": "jammit 개발을 위한 API 문서입니다.\n\n🔗 OpenAPI 명세 다운로드: /openapi-spec.json",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "JWT access token",
        "in": "header"
      }
    },
    "schemas": {
...
      "VideoListDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "b1a8f3e1-3b0f-4e9b-98a2-c4f0e6d3a3b4",
            "description": "영상 UUID"
          },
          "title": {
            "type": "string",
            "example": "재미있게 합주한 아이스크림",
            "description": "비디오 제목"
          },
          "thumbnailUrl": {
            "type": "string",
            "example": "https://picsum.photos/200/300",
            "description": "썸네일 이미지 URL"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2023-05-09T12:34:56Z",
            "description": "생성일"
          },
          "nickname": {
            "type": "string",
            "example": "퉁퉁퉁퉁퉁퉁사후르",
            "description": "사용자 닉네임"
          },
          "duration": {
            "type": "string",
            "example": "00:00:30",
            "description": "비디오 길이"
          }
        },
        "required": [
          "id",
          "title",
          "thumbnailUrl",
          "createdAt",
          "nickname",
          "duration"
        ]
      },
      "VideoListFlatDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number"
          },
          "take": {
            "type": "number"
          },
          "total": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoListDto"
            }
          }
        },
        "required": [
          "page",
          "take",
          "total",
          "totalPages",
          "data"
        ]
      },
      "VideoResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "b1a8f3e1-3b0f-4e9b-98a2-c4f0e6d3a3b4",
            "description": "영상 UUID"
          },
          "title": {
            "type": "string",
            "example": "재미있게 합주한 아이스크림",
            "description": "비디오 제목"
          },
          "thumbnailUrl": {
            "type": "string",
            "example": "https://picsum.photos/200/300",
            "description": "썸네일 이미지 URL"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2023-05-09T12:34:56Z",
            "description": "생성일"
          },
          "nickname": {
            "type": "string",
            "example": "퉁퉁퉁퉁퉁퉁사후르",
            "description": "사용자 닉네임"
          },
          "duration": {
            "type": "string",
            "example": "00:00:30",
            "description": "비디오 길이"
          },
          "description": {
            "type": "string",
            "example": "다같이 아이스크림 췃어요",
            "description": "비디오 설명"
          },
          "playbackId": {
            "type": "string",
            "example": "C52Fu8r79UfIp1GX02jE8lnLBE4th5E2TwC7YfTaYnwU",
            "description": "비디오 재생 URL"
          },
          "viewCount": {
            "type": "number",
            "example": 103,
            "description": "조회수"
          },
          "likeCount": {
            "type": "number",
            "example": 1,
            "description": "유저 고유아이디"
          },
          "commentCount": {
            "type": "number",
            "example": 4,
            "description": "댓글 수"
          }
        },
        "required": [
          "id",
          "title",
          "thumbnailUrl",
          "createdAt",
          "nickname",
          "duration",
          "description",
          "playbackId",
          "viewCount",
          "likeCount",
          "commentCount"
        ]
      },
      "MuxUploadResponseDto": {
        "type": "object",
        "properties": {
          "uploadUrl": {
            "type": "string",
            "example": "https://storage.mux.com/abc123...",
            "description": "프론트에서 사용할 Direct Upload URL"
          },
          "uploadId": {
            "type": "string",
            "example": "upload-xyz456",
            "description": "Mux 업로드 식별자 ID (uploadId)"
          }
        },
        "required": [
          "uploadUrl",
          "uploadId"
        ]
      },
      "RegisterMuxVideoDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "재미있게 합주한 아이스크림",
            "description": "비디오 제목"
          },
          "description": {
            "type": "string",
            "example": "다같이 아이스크림 췃어요",
            "description": "비디오 설명"
          },
          "uploadId": {
            "type": "string",
            "example": "upload-xyz456",
            "description": "업로드 아이디"
          },
          "thumbnailUrl": {
            "type": "string",
            "example": "https://picsum.photos/200/300",
            "description": "썸네일 이미지 URL"
          },
          "slug": {
            "type": "number",
            "example": 1,
            "description": "상세페이지 아이디"
          }
        },
        "required": [
          "title",
          "description",
          "uploadId",
          "thumbnailUrl",
          "slug"
        ]
      },
...
    }
  }
}